+++ backend/canisters/community/impl/src/memory.rs
@@ -8,6 +8,7 @@ const UPGRADES: MemoryId = MemoryId::new(0);
 const INSTRUCTION_COUNTS_INDEX: MemoryId = MemoryId::new(1);
 const INSTRUCTION_COUNTS_DATA: MemoryId = MemoryId::new(2);
 const STABLE_MEMORY_MAP: MemoryId = MemoryId::new(3);
+const STABLE_MEMORY_MAP_SMALL_ENTRIES: MemoryId = MemoryId::new(4);
 
 pub type Memory = VirtualMemory<DefaultMemoryImpl>;
 
@@ -32,8 +33,12 @@ pub fn get_stable_memory_map_memory() -> Memory {
     get_memory(STABLE_MEMORY_MAP)
 }
 
+pub fn get_stable_memory_map_small_entries_memory() -> Memory {
+    get_memory(STABLE_MEMORY_MAP_SMALL_ENTRIES)
+}
+
 pub fn memory_sizes() -> BTreeMap<u8, u64> {
-    (0u8..=3).map(|id| (id, get_memory(MemoryId::new(id)).size())).collect()
+    (0u8..=4).map(|id| (id, get_memory(MemoryId::new(id)).size())).collect()
 }
 
 fn get_memory(id: MemoryId) -> Memory {